Move shift_book endpoints to ShiftController
