+++ app/Resources/views/booking/index.html.twig
@@ -190,7 +190,7 @@
 
             if (selectedFormation !== null && selectedTypeService !== null) {
 
-                // let url = "/booking/shift/" + checkedValue +"/book";
+                // let url = "/shift/" + checkedValue +"/book";
 
                 let url = "{{ path('shift_book',{'id': 'tmpId'}) }}";
                 url = url.replace("tmpId", selectedFormation.value);
