+++ src/AppBundle/Controller/BookingController.php
@@ -171,7 +171,7 @@ class BookingController extends Controller
 
     /**
      * Build the filter form for the admin main page (route /booking/admin)
-     * and rerun an array with  the form object and the date range and the action
+     * and rerun an array with the form object and the date range and the action
      *
      * the return object :
      * array(
@@ -385,7 +385,6 @@ class BookingController extends Controller
      */
     public function adminAction(Request $request): Response
     {
-
         $filter = $this->adminFilterFormFactory($request);
 
         // calendar creation
@@ -415,15 +414,14 @@ class BookingController extends Controller
      */
     public function showBucketAction(Request $request, Shift $bucket)
     {
-
         $em = $this->getDoctrine()->getManager();
         $shifts = $em->getRepository('AppBundle:Shift')->findBucket($bucket);
 
-        $shift_book_forms = [];
+        $shiftBookForms = [];
         foreach ($shifts as $shift) {
-            $shift_book_forms[$shift->getId()] = $this->createBookForm($shift)->createView();
+            $shiftBookForms[$shift->getId()] = $this->createBookForm($shift)->createView();
         }
-        $shift_add_form = $this->createForm(
+        $shiftAddForm = $this->createForm(
             ShiftType::class,
             $bucket,
             array(
@@ -435,8 +433,8 @@ class BookingController extends Controller
 
         return $this->render('admin/booking/_partial/bucket_modal.html.twig', [
             'shifts' => $shifts,
-            'shift_add_form' => $shift_add_form,
-            'shift_book_forms' => $shift_book_forms
+            'shift_add_form' => $shiftAddForm,
+            'shift_book_forms' => $shiftBookForms
         ]);
 
     }
@@ -524,65 +522,7 @@ class BookingController extends Controller
         return $this->createFormBuilder()
             ->setAction($this->generateUrl('delete_bucket', array('id' => $bucket->getId())))
             ->setMethod('DELETE')
-            ->getForm()
-        ;
-    }
-
-    /**
-     * Book a shift.
-     *
-     * @Route("/shift/{id}/book", name="shift_book")
-     * @Method("POST")
-     */
-    public function bookShiftAction(Shift $shift,Request $request): Response
-    {
-        $session = new Session();
-
-        $content = json_decode($request->getContent());
-        $beneficiaryId = $content->beneficiaryId;
-        $isFixe = $content->typeService;
-
-        $em = $this->getDoctrine()->getManager();
-        $beneficiary = $em->getRepository('AppBundle:Beneficiary')->find($beneficiaryId);
-
-        // Check if the shift is bookable by the given beneficiary
-        // Also check if the beneficiary belongs to the same membership as the current user
-        if (!$beneficiary
-            || !$this->get('shift_service')->isShiftBookable($shift, $beneficiary)
-            || !$this->isGranted(MembershipVoter::EDIT, $beneficiary->getMembership())
-        ) {
-            $session->getFlashBag()->add("error", "Impossible de réserver ce créneau");
-            return new Response($this->generateUrl('booking'), 205);
-        }
-
-        if (!$shift->getBooker()) {
-            $current_user = $this->get('security.token_storage')->getToken()->getUser();
-            $shift->setBooker($current_user);
-            $shift->setBookedTime(new DateTime('now'));
-        }
-        $shift->setShifter($beneficiary);
-        $shift->setIsDismissed(false);
-        $shift->setDismissedReason(null);
-        $shift->setDismissedTime(null);
-        $shift->setLastShifter(null);
-        $shift->setFixe($isFixe);
-        $em->persist($shift);
-
-        $member = $beneficiary->getMembership();
-        if ($member->getFirstShiftDate() == null) {
-            $firstDate = clone($shift->getStart());
-            $firstDate->setTime(0, 0, 0);
-            $member->setFirstShiftDate($firstDate);
-            $em->persist($member);
-        }
-
-        $em->flush();
-
-        $dispatcher = $this->get('event_dispatcher');
-        $dispatcher->dispatch(ShiftBookedEvent::NAME, new ShiftBookedEvent($shift, false));
-
-        $session->getFlashBag()->add("success", "Ce créneau a bien été réservé !");
-        return new Response($this->generateUrl('homepage'), 200);
+            ->getForm();
     }
 
     /**
@@ -735,68 +675,6 @@ class BookingController extends Controller
         }
 
         return $this->redirectToRoute('homepage');
-    }
-
-    /**
-     * Book a shift admin.
-     *
-     * @Route("/admin/shift/{id}/book", name="admin_shift_book")
-     * @Security("has_role('ROLE_SHIFT_MANAGER')")
-     * @Method("POST")
-     */
-    public function bookShiftAdminAction(Request $request, Shift $shift)
-    {
-        $session = new Session();
-
-        $form = $this->createBookForm($shift);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            if ($shift->getShifter() && !$shift->getIsDismissed()) {
-                $session->getFlashBag()->add("error", "Désolé, ce créneau est déjà réservé");
-                return $this->redirectToRoute('booking_admin');
-            }
-
-            $fixe = $form->get("fixe")->getData();
-            $beneficiary = $form->get("shifter")->getData();
-
-            if ($shift->getFormation() && !$beneficiary->getFormations()->contains($shift->getFormation())) {
-                $session->getFlashBag()->add("error", "Désolé, ce bénévole n'a pas la qualification necessaire (" . $shift->getFormation()->getName() . ")");
-                return $this->redirectToRoute('booking_admin');
-            }
-
-            if (!$shift->getBooker()) {
-                $current_user = $this->get('security.token_storage')->getToken()->getUser();
-                $shift->setBooker($current_user);
-                $shift->setBookedTime(new DateTime('now'));
-            }
-            $shift->setShifter($beneficiary);
-            $shift->setIsDismissed(false);
-            $shift->setDismissedReason(null);
-            $shift->setDismissedTime(null);
-            $shift->setLastShifter(null);
-            $shift->setFixe($fixe);
-
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($shift);
-
-            $member = $beneficiary->getMembership();
-            if ($member->getFirstShiftDate() == null) {
-                $firstDate = clone($shift->getStart());
-                $firstDate->setTime(0, 0, 0);
-                $member->setFirstShiftDate($firstDate);
-                $em->persist($member);
-            }
-            $em->flush();
-
-            $dispatcher = $this->get('event_dispatcher');
-            $dispatcher->dispatch(ShiftBookedEvent::NAME, new ShiftBookedEvent($shift, true));
-
-            $session->getFlashBag()->add("success", "Créneau réservé avec succès pour " . $shift->getShifter());
-            return $this->redirectToRoute('booking_admin');
-        }
-        $session->getFlashBag()->add('error', "Une erreur est survenue...");
-        return $this->redirectToRoute('booking_admin');
     }
 
     /**
@@ -964,8 +842,8 @@ class BookingController extends Controller
     private function createBookForm(Shift $shift)
     {
         $form = $this->get('form.factory')->createNamedBuilder('shift_book_forms_' . $shift->getId())
-            ->setAction($this->generateUrl('admin_shift_book', array('id' => $shift->getId())))
-            ->add('shifter', AutocompleteBeneficiaryType::class, array('label'=>'Numéro d\'adhérent ou nom du membre', 'required'=>true));
+            ->setAction($this->generateUrl('shift_book_admin', array('id' => $shift->getId())))
+            ->add('shifter', AutocompleteBeneficiaryType::class, array('label' => 'Numéro d\'adhérent ou nom du membre', 'required' => true));
 
         if ($this->useFlyAndFixed) {
             $form = $form->add('fixe', RadioChoiceType::class, [
@@ -983,5 +861,4 @@ class BookingController extends Controller
 
         return $form->getForm();
     }
-
 }
